Split movement into generic and player-based
